# vlife-admin/src/main/java/cn/wwwlike/oa/dto/ProjectPinDto.java
/*
 *  vlife http://github.com/wwwlike/vlife
 *
 *  Copyright (C)  2018-2022 vlife
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 */

package cn.wwwlike.oa.dto;

import cn.wwwlike.oa.entity.Project;
import cn.wwwlike.vlife.base.SaveBean;
import lombok.Data;

/**
 * 收藏
 *
 * @author xiaoyu
 * @date 2022/7/5
 */
@Data
public class ProjectPinDto implements SaveBean<Project> {
    /**
     * 项目id
     */
    public String id;
    /**
     * 是否收藏
     */
    public boolean pin;
}