# vlife-admin/src/main/resources/static/doc/openapi.json
{
  "components": {
    "schemas": {
      "ProjectDto": {
        "properties": {
          "id": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "accountId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "Task": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "projectId": {
            "format": "string",
            "description": "所属项目",
            "type": "string"
          },
          "point": {
            "format": "int32",
            "description": "任务点数",
            "type": "integer"
          },
          "progress": {
            "format": "double",
            "description": "当前完成进度",
            "type": "number"
          },
          "state": {
            "format": "string",
            "description": "任务完成状态",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "任务名称",
            "type": "string"
          },
          "sort": {
            "format": "int32",
            "description": "排序",
            "type": "integer"
          },
          "level": {
            "format": "string",
            "description": "紧急程度(字典)",
            "type": "string"
          },
          "stepId": {
            "format": "string",
            "description": "当前所在阶段",
            "type": "string"
          },
          "accountId": {
            "format": "string",
            "description": "任务创建人",
            "type": "string"
          }
        }
      },
...
      "Dept": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "部门名称",
            "type": "string"
          }
        }
      },
...
      "Step": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          }
        }
      },
...
      "ProjectAccount": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "projectId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "accountId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "PageVoProject": {
        "properties": {
          "result": {
            "description": "No comments found.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "size": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "page": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "totalPage": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "first": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          },
          "last": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          }
        }
      },
...
      "ProjectPinDto": {
        "properties": {
          "id": {
            "format": "string",
            "description": "项目id",
            "type": "string"
          },
          "pin": {
            "format": "boolean",
            "description": "是否收藏",
            "type": "boolean"
          }
        }
      },
...
      "AccountVo": {
        "properties": {
          "id": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "token": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "password": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "deptId": {
            "format": "string",
            "description": "用户组",
            "type": "string"
          }
        }
      },
...
      "Project": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "项目名称",
            "type": "string"
          },
          "pin": {
            "format": "boolean",
            "description": "收藏\r\n暂时所有人公用",
            "type": "boolean"
          },
          "accountId": {
            "format": "string",
            "description": "项目负责人",
            "type": "string"
          },
          "planDate": {
            "format": "string",
            "description": "计划开始日期",
            "type": "string"
          },
          "finishedDate": {
            "format": "string",
            "description": "结束日期",
            "type": "string"
          },
          "state": {
            "format": "string",
            "description": "项目状态",
            "type": "string"
          },
          "type": {
            "format": "string",
            "description": "项目类型",
            "type": "string"
          },
          "totalPoint": {
            "format": "int32",
            "description": "项目任务总点数",
            "type": "integer"
          },
          "finishedPoint": {
            "format": "int32",
            "description": "已经完成的任务点数",
            "type": "integer"
          }
        }
      },
...
      "ProjectDetailVo": {
        "properties": {
          "id": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "项目名称",
            "type": "string"
          },
          "account_name": {
            "format": "string",
            "description": "项目负责人姓名",
            "type": "string"
          },
          "planDate": {
            "format": "string",
            "description": "计划开始日期",
            "type": "string"
          },
          "finishedDate": {
            "format": "string",
            "description": "结束日期",
            "type": "string"
          },
          "state": {
            "format": "string",
            "description": "项目状态 project_state",
            "type": "string"
          },
          "type": {
            "format": "string",
            "description": "项目类型",
            "type": "string"
          },
          "totalPoint": {
            "format": "int32",
            "description": "项目任务总点数",
            "type": "integer"
          },
          "finishedPoint": {
            "format": "int32",
            "description": "已经完成的任务点数",
            "type": "integer"
          },
          "steps": {
            "description": "项目的生命周期",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Step"
            }
          },
          "tasks": {
            "description": "项目下的任务",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          }
        }
      },
...
      "AccountDto": {
        "properties": {
          "id": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "username": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "password": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "group": {
            "description": "No comments found.(object)",
            "type": "object",
            "$ref": "#/components/schemas/Dept"
          },
          "project_id": {
            "format": "string",
            "description": "创建人的id",
            "type": "string"
          },
          "project_name": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "projects": {
            "description": "项目创建人",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "projectAccount_projectId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "projectAccount_project": {
            "description": "多对多的关联关系",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "projectAccount_project_name": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8088"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
...
    "/project/page": {
      "get": {
        "summary": "分页查询项目表;",
        "deprecated": false,
        "description": "分页查询项目表;",
        "tags": [
          "项目表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageVoProject"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"result\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"gtwgw6\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"name\": \"ellie.kuphal\",\n      \"pin\": true,\n      \"accountId\": \"144\",\n      \"planDate\": \"2022-08-13\",\n      \"finishedDate\": \"2022-08-13\",\n      \"state\": \"vhffen\",\n      \"type\": \"8ppef5\",\n      \"totalPoint\": 165,\n      \"finishedPoint\": 179\n    }\n  ],\n  \"size\": 10,\n  \"page\": 1,\n  \"total\": 786,\n  \"totalPage\": 994,\n  \"first\": true,\n  \"last\": true\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "page",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": "144",
            "name": "account_deptId"
          },
          {
            "schema": {
              "type": "array",
              "items": {}
            },
            "in": "query",
            "description": "项目类型",
            "required": false,
            "example": "uboqv9,90l2nk",
            "name": "type"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "根据负责人id查询",
            "required": false,
            "example": "144",
            "name": "accountId"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "根据参与人Id查询(默认都需要传入)",
            "required": false,
            "example": "144",
            "name": "projectAccount_accountId"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "项目状态",
            "required": false,
            "example": "aom12m",
            "name": "state"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "项目名称",
            "required": false,
            "example": "ellie.kuphal",
            "name": "name"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "项目开始年度",
            "required": false,
            "example": "860",
            "name": "year"
          }
        ]
      }
    },
...
    "/project/detail/{id}": {
      "get": {
        "summary": "明细查询类说明;",
        "deprecated": false,
        "description": "明细查询类说明;",
        "tags": [
          "项目表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailVo"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"name\": \"ellie.kuphal\",\n  \"account_name\": \"ellie.kuphal\",\n  \"planDate\": \"2022-08-13\",\n  \"finishedDate\": \"2022-08-13\",\n  \"state\": \"ai7vkf\",\n  \"type\": \"uxhpr2\",\n  \"totalPoint\": 244,\n  \"finishedPoint\": 526,\n  \"steps\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"n71t4f\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\"\n    }\n  ],\n  \"tasks\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"bjtpye\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"projectId\": \"144\",\n      \"point\": 715,\n      \"progress\": 44.70,\n      \"state\": \"3l9a3j\",\n      \"name\": \"ellie.kuphal\",\n      \"sort\": 418,\n      \"level\": \"rp0mi7\",\n      \"stepId\": \"144\",\n      \"accountId\": \"144\"\n    }\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/projectAccount/detail/{id}": {
      "get": {
        "summary": "明细查询项目人员关系表(分配关系);",
        "deprecated": false,
        "description": "明细查询项目人员关系表(分配关系);",
        "tags": [
          "项目人员关系表(分配关系)接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectAccount"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"fj8ggz\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"projectId\": \"144\",\n  \"accountId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/account/detail/{id}": {
      "get": {
        "summary": "明细查询类说明;",
        "deprecated": false,
        "description": "明细查询类说明;",
        "tags": [
          "员工用户接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountVo"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"name\": \"ellie.kuphal\",\n  \"token\": \"ms3d71\",\n  \"password\": \"hroym7\",\n  \"deptId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/project/save": {
      "post": {
        "summary": "保存项目编辑;",
        "deprecated": false,
        "description": "保存项目编辑;",
        "tags": [
          "项目表接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDto"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"name\": \"ellie.kuphal\",\n  \"accountId\": \"144\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"name\": \"ellie.kuphal\",\n  \"accountId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/projectAccount/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "项目人员关系表(分配关系)接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "299"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/projectAccount/save": {
      "post": {
        "summary": "保存项目人员关系表(分配关系);",
        "deprecated": false,
        "description": "保存项目人员关系表(分配关系);",
        "tags": [
          "项目人员关系表(分配关系)接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectAccount"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"status\": \"roh6wd\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"projectId\": \"144\",\n  \"accountId\": \"144\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectAccount"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"t81e2d\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"projectId\": \"144\",\n  \"accountId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/project/save/pin": {
      "post": {
        "summary": "保存收藏;",
        "deprecated": false,
        "description": "保存收藏;",
        "tags": [
          "项目表接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPinDto"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"pin\": true\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPinDto"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"pin\": true\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "savePin",
        "parameters": []
      }
    },
...
    "/account/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "员工用户接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "998"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
    "/account/save": {
      "post": {
        "summary": "保存登录元素;",
        "deprecated": false,
        "description": "保存登录元素;",
        "tags": [
          "员工用户接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDto"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"username\": \"ellie.kuphal\",\n  \"password\": \"extsb3\",\n  \"group\": {\n    \"id\": \"144\",\n    \"status\": \"t2a505\",\n    \"createDate\": \"2022-08-13\",\n    \"modifyDate\": \"2022-08-13\",\n    \"name\": \"ellie.kuphal\"\n  },\n  \"project_id\": [\n    \"azijts\"\n  ],\n  \"project_name\": [\n    \"uuvqr9\"\n  ],\n  \"projects\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"ufzlg4\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"name\": \"ellie.kuphal\",\n      \"pin\": true,\n      \"accountId\": \"144\",\n      \"planDate\": \"2022-08-13\",\n      \"finishedDate\": \"2022-08-13\",\n      \"state\": \"v0ln68\",\n      \"type\": \"8ucmpk\",\n      \"totalPoint\": 940,\n      \"finishedPoint\": 923\n    }\n  ],\n  \"projectAccount_projectId\": [\n    \"h9dl9l\"\n  ],\n  \"projectAccount_project\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"x8su1b\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"name\": \"ellie.kuphal\",\n      \"pin\": true,\n      \"accountId\": \"144\",\n      \"planDate\": \"2022-08-13\",\n      \"finishedDate\": \"2022-08-13\",\n      \"state\": \"l1xcek\",\n      \"type\": \"k9crbh\",\n      \"totalPoint\": 309,\n      \"finishedPoint\": 251\n    }\n  ],\n  \"projectAccount_project_name\": [\n    \"e7ln2z\"\n  ]\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"username\": \"ellie.kuphal\",\n  \"password\": \"url1nv\",\n  \"group\": {\n    \"id\": \"144\",\n    \"status\": \"v9ygl5\",\n    \"createDate\": \"2022-08-13\",\n    \"modifyDate\": \"2022-08-13\",\n    \"name\": \"ellie.kuphal\"\n  },\n  \"project_id\": [\n    \"q6g64i\"\n  ],\n  \"project_name\": [\n    \"01u7rf\"\n  ],\n  \"projects\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"jcldkf\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"name\": \"ellie.kuphal\",\n      \"pin\": true,\n      \"accountId\": \"144\",\n      \"planDate\": \"2022-08-13\",\n      \"finishedDate\": \"2022-08-13\",\n      \"state\": \"xqaye3\",\n      \"type\": \"uap0ft\",\n      \"totalPoint\": 326,\n      \"finishedPoint\": 719\n    }\n  ],\n  \"projectAccount_projectId\": [\n    \"a948kc\"\n  ],\n  \"projectAccount_project\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"va3n05\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"name\": \"ellie.kuphal\",\n      \"pin\": true,\n      \"accountId\": \"144\",\n      \"planDate\": \"2022-08-13\",\n      \"finishedDate\": \"2022-08-13\",\n      \"state\": \"q3k0xq\",\n      \"type\": \"vrejwb\",\n      \"totalPoint\": 593,\n      \"finishedPoint\": 12\n    }\n  ],\n  \"projectAccount_project_name\": [\n    \"32k70k\"\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/project/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "项目表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "275"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "test",
    "version": "1.0.0"
  },
...
}